"Updated footer.blade.php: added total slide count to listingDetailsSliderNav"
This commit updates the footer.blade.php file by adding the total number of slides to the listingDetailsSliderNav. When a user hovers over a slide in the navigation, the updated code displays the current slide number and the total number of slides in the format "current slide / total slides". This provides a better user experience by giving them a clearer understanding of the number of slides in the slider.