"Refactor filtering logic in HomeController and ProjectController; update filter rendering in menu-list.blade.php"
This commit includes changes to the HomeController and ProjectController files, as well as the menu-list.blade.php view. The filtering logic has been refactored, and some conditionals have been updated in both controllers. In the view, filter rendering has been updated based on the changes in the controllers. The commit does not introduce new functionality but aims to improve the existing code structure and readability.