"Updated project-item-card view: modified display of projectHousingsList data based on additional condition"
This commit message describes the changes in the `project-item-card.blade.php` file. Specifically, it mentions that the display of `projectHousingsList` data has been updated based on an additional condition in the `$column_additional` variable. The condition now checks if `$column_additional` is not only set but also numeric. This change ensures that the additional information is displayed correctly for each project housing item.