Update cart view to check if `$share_sale` is set before displaying
The view file for the shopping cart index page has been updated to check if the variable `$share_sale` is set before displaying its value. If `$share_sale` is not set, the view will now display a null value instead of throwing an error. This change ensures that the view is more resilient to changes in the data that it receives.