Update `project-item-card.blade.php`: Remove `print_r` statement and adjust conditional statement for displaying "Off Sale" button
This commit removes an unused `print_r` statement for debugging purposes and modifies the conditional statement for displaying the "Off Sale" button based on the value of `$projectHousingsList[$keyIndex]['off_sale[]']`. The "Off Sale" button will only be displayed if `$projectHousingsList[$keyIndex]['off_sale[]']` is not equal to an empty array and if the property is not already sold.