"Add Open Graph meta tags to header for better social media sharing"
This commit adds Open Graph meta tags to the header of the page. These tags include information such as the site name, URL, type, title, description, and image, which will be used by social media platforms when the page is shared. The title and description are dynamically set based on the `$pageInfo` variable. The image is set to a small logo image.