"Changed CartOrder query to exclude disabled orders in ProfileController"
This commit message accurately describes the code diff by specifying that the `CartOrder` query in the `ProfileController` has been modified to exclude orders with a status of "disabled". The message is concise and sticks to the facts of the code change without making any assumptions or interpretations about why the change was made.