"Add flag to track if cart has counter and update cart item query in add method of CartController"
This commit updates the `add` method in the `CartController` by adding a new flag `$hasCounter` and modifying the query for retrieving the latest cart item. The purpose of these changes is to track if the cart has a counter and to optimize the query for fetching the most recent cart item.