Add link to delete neighbor for sold housing units in admin panel
This commit adds a link to delete a neighbor for sold housing units in the admin panel. The link is added next to the "Sipariş Detayı" (Order Details) link and is only visible if the housing unit has been sold. The link leads to a confirmation page for deleting the neighbor. The neighbor data will be deleted from the database when the link is clicked. The link has a red background and the text "Komşumu Sil" (Delete Neighbor).