Added new menu item for "Gelir Tablosu" (Income Table) in admin menu
This commit adds a new item to the admin menu, labeled "Muhasebe" (which means "accounting" in Turkish), with an icon of a dollar sign. The new item has a text label of "Gelir Tablosu" (which means "income table" in Turkish) and is associated with the key "ViewAccounting". The new item points to the URL "admin.accounting" and has an active path of "qR9zLp2xS6y/secured/accounting".