"Update HomeController to use constructor injection for SmsService and add sendSMS method to send SMS notification for refund status updates"
This commit updates the HomeController class by using constructor injection for the SmsService. It also adds a new method called sendSMS to send SMS notifications when the refund status is updated. The SmsService is used to send the SMS message to the user's phone number. The commit also modifies the upload method to use the new sendSMS method to send an SMS notification when a refund receipt is uploaded.