"Updated project-item-card view: added debug output for $sold variable"
This commit updates the `resources/views/components/project-item-card.blade.php` file by adding the `{{print_r($sold)}}` debug output in the payment button section. The `$sold` variable is printed in addition to the existing `$sold_check` variable. This change allows for the display of the `$sold` variable's value during development or debugging.