Update project-item-card component: Replace `{{$dataSoldValue}}` with condition for displaying "Off Sale" badge
The diff updates the `resources/views/components/project-item-card.blade.php` file by replacing the line `{{$dataSoldValue}}` with a new condition in the payment button. The new condition checks if the project is not sold and if the `off_sale` property of the corresponding housing in the `$projectHousingsList` array is set and not empty. If the condition is true, the text "Off Sale" will be displayed instead of the `{{$dataSoldValue}}`.