"Updated data attribute value for payment button in project-item-card component"
This commit updates the data attribute value for the payment button in the project-item-card component. The line of code that was changed is related to the display of the payment button based on certain conditions. The old line of code checked if the project was not sold and if there were any off-sale housing units available. The new line of code checks if there is a value in the $sumCartOrderQt array for the current index and if the value of the 'qt_total' key is equal to the number of shares.