Update $dataSoldValue calculation in project-item-card component
The calculation of the $dataSoldValue variable has been updated in the project-item-card component. The condition related to the $share\_sale variable has been moved outside of the first condition. Additionally, the check for $sumCartOrderQt[$keyIndex]['qt\_total'] == $number\_of\_share is now only checked once.

The button for the payment plan button now displays "1" if the project order is already in the cart and the number of shares matches, or if the project order is sold and the share sale is not defined or empty. It displays "0" otherwise.