Update project-item-card component: add ternary expression to payment button
The changes include adding a ternary expression to the payment button's data-payment-order attribute. The expression checks if the project has been sold, if the sale status is not 2, if the share sale is empty or equal to an empty array, if the sum of the cart order quantity is equal to the number of shares, or if the project housing is off-sale and not empty. If any of these conditions are met, the data-payment-order attribute will be set to 1, otherwise, it will remain unchanged.