"Update data-sold attribute in project-item-card and project-item-mobile-card components"
This commit updates the `data-sold` attribute in the `project-item-card` and `project-item-mobile-card` components. It simplifies the expression used to calculate the value of the attribute and stores it in a separate PHP variable `$dataSoldValue`. The new expression checks if the project is sold, if the share sale is empty or not, if the project has any off-sale items, and if the number of shares in the cart matches the number of shares in the project. The `$dataSoldValue` is then used to set the value of the `data-sold` attribute.