Update `data-sold` attribute in project-item-mobile-card component
The `data-sold` attribute in the "payment-plan-mobile-btn" button now checks if `$sold->status` is not equal to 2 instead of checking if it's equal to 1 or 0. This change ensures that the button is not displayed if the project is already sold out, regardless of the status value. The rest of the conditions for displaying the button remain unchanged.