"Added ternary expression output to payment detail button in project-item-card component"
This commit updates the "resources/views/components/project-item-card.blade.php" file by adding a ternary expression output to the payment detail button. The output displays the value of the ternary expression, which is either 1 or 0, depending on the status of the sold property, its status, and the share sale availability.