"Updated project-item-card.blade.php: Changed conditionals for displaying payment details button"
The diff includes changes to the project-item-card.blade.php file where conditionals for displaying the payment details button have been updated. Specifically, the code now checks if the variable $share\_sale\_empty is not empty before checking if the sum of the cart order quantity is equal to the number of shares. This change was made in three different places in the code.