"Add 'see my neighbor' button for sold properties with neighbor view status 2 in project-item-card and project-item-mobile-card components"
This commit adds a new button for sold properties that have a neighbor view status of 2. The button is only displayed if the user is logged in and has access to view the neighbor. If the user is not logged in, clicking the button will redirect them to the login page. The button is added in two different blade templates, project-item-card.blade.php and project-item-mobile-card.blade.php. The functionality of the button is the same in both templates.