"Updated neighbor project selection in payment index view to use array syntax for accessing owner properties"
This commit message describes the changes in the code diff for the payment index view in a blade template. The changes include updating the syntax for accessing owner properties of neighbor projects from object notation to array notation. The commit message does not make any interpretations or assumptions about why the changes were made, but simply describes the differences between the two versions of the code.