"Update invoice index view: adjust project information display based on project and housing data"
This commit updates the 'resources/views/admin/invoice/index.blade.php' file to modify the display of project information in the invoice index view. It introduces conditional checks to handle cases where the project data is an instance of the \App\Models\Housing class and also considers the 'isShare' property of the housing data. This results in more accurate and relevant project information being displayed in the invoice index.