Invoices: Change condition for displaying share information
The code diff modifies the view files for the admin and institutional invoices. It changes the condition for displaying share information in the invoice details. The original condition checked if `$cart['item']['isShare']` was greater than 0. This has been changed to check if the count of `$cart['item']['isShare']` is greater than 0. This change affects how share information is displayed in the invoice.